// src/_services/api/model/transporteResponseZodDtoOutput.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Produtividade API
 * API de produtividade para gestão de produtividade dos funcionários dentro do armazém e gestão de carregamento
 * OpenAPI spec version: 1.0
 */
import type { TransporteResponseZodDtoOutputStatus } from './transporteResponseZodDtoOutputStatus';
import type { TransporteResponseZodDtoOutputSeparacao } from './transporteResponseZodDtoOutputSeparacao';
import type { TransporteResponseZodDtoOutputConferencia } from './transporteResponseZodDtoOutputConferencia';
import type { TransporteResponseZodDtoOutputCarregamento } from './transporteResponseZodDtoOutputCarregamento';

export interface TransporteResponseZodDtoOutput {
  numeroTransporte: string;
  status: TransporteResponseZodDtoOutputStatus;
  nomeRota: string;
  nomeTransportadora: string;
  placa: string;
  temCortes?: boolean;
  cadastradoPorId?: string;
  prioridade?: number;
  obs?: string;
  qtdImpressaoSeparacao?: number;
  qtdImpressaoCarregamento?: number;
  separacao: TransporteResponseZodDtoOutputSeparacao;
  conferencia: TransporteResponseZodDtoOutputConferencia;
  carregamento: TransporteResponseZodDtoOutputCarregamento;
}